+++ CHANGES.md
@@ -1,5 +1,6 @@
 # 2.9.0
 
+* Added log target support to `ops.pebble` layers and plans.
 * Added `Harness.run_action()`, `testing.ActionOutput`, and `testing.ActionFailed`
 
 # 2.8.0